Remove RSG from the components
