+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 2.0.0
 
-[The component `RSGButton` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGButton)
+[The component `Button` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGButton)
 
 ### 2.0.1
 
@@ -12,16 +12,16 @@ README.md added.
 
 ### 2.0.2
 
-[The component `RSGBox` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGBox)
+[The component `Box` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGBox)
 
 ### 2.0.3
 
-- [The component `RSGLabel` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGLabel)
-- `RSGBox` bug fixed.
+- [The component `Label` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGLabel)
+- `Box` bug fixed.
 
 ### 2.0.4
 
-`RSGLabel` bug fix.
+`Label` bug fix.
 
 ### 2.0.5
 
@@ -29,16 +29,16 @@ Added GitHub repository in `package.json`
 
 ### 2.0.6
 
-[The component `RSGFormBasic` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGFormBasic)
+[The component `FormBasic` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGFormBasic)
 
 ### 2.0.7
 
-- [The `RSGTransition` style function has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGTransition)
+- [The `Transition` style function has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGTransition)
 - Also added sample paragraph in index.html
 
 ### 2.0.8
 
-- [The component `RSGProgressBar` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGProgressBar)
+- [The component `ProgressBar` has been introduced.](https://github.com/RSG-Group/rsg-components/wiki/RSGProgressBar)
 - Also added new images in GitHub repo.
 
 ### 2.0.9
@@ -47,13 +47,13 @@ You can now read some tips at `../rsg-components/tips`
 
 ### 2.0.10
 
-- `RSGProgressBar` bug fixed.
+- `ProgressBar` bug fixed.
 - `README.md` edited.
 - Also added new images for wiki.
 
 ### 2.0.11
 
-Update to `RSGProgressBar`
+Update to `ProgressBar`
 
 ### 2.0.12
 
@@ -61,10 +61,14 @@ Minor update to README.md and `images`
 
 ### 2.0.13
 
-A new component, `RSGCheckbox`, has been introduced.
+A new component, `Checkbox`, has been introduced.
 
 ![Checkbox image](https://raw.githubusercontent.com/RSG-Group/rsg-components/master/images/Checkboxes.png)
 
 You can theoretically use this without the `rsg-components` library. [Example here.](http://codepen.io/radicho/pen/jBJjPb)
 
 [More info on its development here.](https://medium.com/@radicho/customized-checkboxes-9190f969b681)
+
+### 3.0.0-alpha.1 - 3.0.0-alpha.5
+
+New code style. Tons of bugfixes. New features and styles.
