+++ less/ProgressBar/animated.less
@@ -1,4 +1,4 @@
-.RSGProgressBar_animated {
+.ProgressBar_animated {
   background: rgb(210, 220, 220);
   box-shadow: inset 0 1.5px 3px rgb(150, 150, 150);
   -webkit-box-shadow: inset 0 1.5px 3px rgb(150, 150, 150);
