+++ less/ProgressBar/checkered.less
@@ -1,4 +1,4 @@
-.RSGProgressBar {
+.ProgressBar_checkered {
   background: rgb(210, 215, 215);
   box-shadow: inset 0 1.5px 3px rgb(150, 150, 150);
   -webkit-box-shadow: inset 0 1.5px 3px rgb(150, 150, 150);
@@ -6,11 +6,10 @@
   height: 20px;
 }
 
-.PP {
+.PP_checkered {
   overflow: hidden;
-  background: rgb(51, 122, 183);
+  background: url(https://raw.githubusercontent.com/RSG-Group/rsg-components/master/images/Progress.png);
   transition: width 1s;
-  width: 1%;
   height: 20px;
   border-radius: 5px;
   text-align: center;
