+++ less/ProgressBar/default.less
@@ -1,4 +1,4 @@
-.RSGProgressBar_checkered {
+.ProgressBar {
   background: rgb(210, 215, 215);
   box-shadow: inset 0 1.5px 3px rgb(150, 150, 150);
   -webkit-box-shadow: inset 0 1.5px 3px rgb(150, 150, 150);
@@ -6,10 +6,11 @@
   height: 20px;
 }
 
-.PP_checkered {
+.PP {
   overflow: hidden;
-  background: url(https://raw.githubusercontent.com/RSG-Group/rsg-components/master/images/Progress.png);
+  background: rgb(51, 122, 183);
   transition: width 1s;
+  width: 1%;
   height: 20px;
   border-radius: 5px;
   text-align: center;
