+++ less/index.less
@@ -1,12 +1,12 @@
 /*  Checkbox CSS code import. */
-@import "RSGCheckbox.less";
+@import "Checkbox.less";
 
 /* ProgressBar CSS code import. */
 // Import default styling.
-@import "RSGProgressBar/default.less";
+@import "ProgressBar/default.less";
 
 // Import animated styling.
-@import "RSGProgressBar/animated.less";
+@import "ProgressBar/animated.less";
 
 // Import checkered styling.
-@import "RSGProgressBar/checkered.less";
+@import "ProgressBar/checkered.less";
